requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
autoflake==2.3.1
bcrypt==4.2.1
black==24.10.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.8
decorator==5.1.1
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
fastapi-limiter==0.1.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.1.3
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
ratelimit==2.2.1
redis==5.2.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
time-machine==2.16.0
tomli==2.1.0
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
win32-setctime==1.1.0
yarl==1.17.2
